+++ packages/web-components/fast-foundation/src/data-grid/data-grid.options.ts
@@ -1,4 +1,6 @@
+import type { SyntheticViewTemplate, ViewTemplate } from "@microsoft/fast-element";
 import type { ValuesOf } from "../utilities/index.js";
+import type { FASTDataGridCell } from "./data-grid-cell.js";
 
 /**
  * Enumerates the data grid auto generated header options
@@ -142,3 +144,72 @@ export const DataGridSelectionBehavior = {
  * @public
  */
 export type DataGridSelectionBehavior = ValuesOf<typeof DataGridSelectionBehavior>;
+
+/**
+ * Defines a column in the grid
+ *
+ * @public
+ */
+export interface ColumnDefinition {
+    /**
+     * Identifies the data item to be displayed in this column
+     * (i.e. how the data item is labelled in each row)
+     */
+    columnDataKey: string;
+
+    /**
+     * Sets the css grid-column property on the cell which controls its placement in
+     * the parent row. If left unset the cells will set this value to match the index
+     * of their column in the parent collection of ColumnDefinitions.
+     */
+    gridColumn?: string;
+
+    /**
+     *  Column title, if not provided columnDataKey is used as title
+     */
+    title?: string;
+
+    /**
+     *  Header cell template
+     */
+    headerCellTemplate?: ViewTemplate | SyntheticViewTemplate;
+
+    /**
+     * Whether the header cell has an internal focus queue
+     */
+    headerCellInternalFocusQueue?: boolean;
+
+    /**
+     * Callback function that returns the element to focus in a custom cell.
+     * When headerCellInternalFocusQueue is false this function is called when the cell is first focused
+     * to immediately move focus to a cell element, for example a cell that is a checkbox could move
+     * focus directly to the checkbox.
+     * When headerCellInternalFocusQueue is true this function is called when the user hits Enter or F2
+     */
+    headerCellFocusTargetCallback?: (cell: FASTDataGridCell) => HTMLElement;
+
+    /**
+     * cell template
+     */
+    cellTemplate?: ViewTemplate | SyntheticViewTemplate;
+
+    /**
+     * Whether the cell has an internal focus queue
+     */
+    cellInternalFocusQueue?: boolean;
+
+    /**
+     * Callback function that returns the element to focus in a custom cell.
+     * When cellInternalFocusQueue is false this function is called when the cell is first focused
+     * to immediately move focus to a cell element, for example a cell that is a checkbox could move
+     * focus directly to the checkbox.
+     * When cellInternalFocusQueue is true this function is called when the user hits Enter or F2
+     */
+
+    cellFocusTargetCallback?: (cell: FASTDataGridCell) => HTMLElement;
+
+    /**
+     * Whether this column is the row header
+     */
+    isRowHeader?: boolean;
+}
